Added descriptive JSDoc comments to methods and components.
