--- app/components/ui/alert.js
@@ -1,49 +1,0 @@
-import * as React from "react";
-import { cva } from "class-variance-authority";
-import { cn } from "../../../lib/utils";
-
-const alertVariants = cva(
-    "relative w-full rounded-lg border px-4 py-3 text-sm [&>svg+div]:translate-y-[-3px] [&>svg]:absolute [&>svg]:left-4 [&>svg]:top-4 [&>svg]:text-foreground [&>svg~*]:pl-7",
-    {
-        variants: {
-            variant: {
-                default: "bg-green-50 text-green-700 border-green-200",
-                destructive:
-                    "border-red-200 bg-red-50 text-red-700 dark:border-red-900",
-            },
-        },
-        defaultVariants: {
-            variant: "default",
-        },
-    }
-);
-
-const Alert = React.forwardRef(({ className, variant, ...props }, ref) => (
-    <div
-        ref={ref}
-        role="alert"
-        className={cn(alertVariants({ variant }), className)}
-        {...props}
-    />
-));
-Alert.displayName = "Alert";
-
-const AlertTitle = React.forwardRef(({ className, ...props }, ref) => (
-    <h5
-        ref={ref}
-        className={cn("mb-1 font-medium leading-none tracking-tight", className)}
-        {...props}
-    />
-));
-AlertTitle.displayName = "AlertTitle";
-
-const AlertDescription = React.forwardRef(({ className, ...props }, ref) => (
-    <div
-        ref={ref}
-        className={cn("text-sm [&_p]:leading-relaxed", className)}
-        {...props}
-    />
-));
-AlertDescription.displayName = "AlertDescription";
-
-export { Alert, AlertTitle, AlertDescription };
