+++ app/components/ui/button.js
@@ -2,6 +2,13 @@ import * as React from "react";
 import { cva } from "class-variance-authority";
 import { cn } from "../../../lib/utils";
 
+/**
+ * Variants for the button component styles.
+ * @param {Object} props - The component props.
+ * @param {string} props.variant - The style variant of the button.
+ * @param {string} props.size - The size variant of the button.
+ * @returns {string} The CSS class for the button component.
+ */
 const buttonVariants = cva(
     "inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium transition-colors focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-green-700 disabled:pointer-events-none disabled:opacity-50",
     {
@@ -32,6 +39,16 @@ const buttonVariants = cva(
     }
 );
 
+/**
+ * Button component for various use cases.
+ * @param {Object} props - The component props.
+ * @param {string} [props.className] - Additional CSS classes to apply to the button.
+ * @param {string} [props.variant] - The style variant of the button.
+ * @param {string} [props.size] - The size variant of the button.
+ * @param {boolean} [props.asChild] - Whether the button should render as another component.
+ * @returns {JSX.Element} The rendered button component.
+ * @param {...Object} props.rest - Additional props for the button.
+ */
 const Button = React.forwardRef(
     ({ className, variant, size, asChild = false, ...props }, ref) => {
         const Comp = asChild ? Slot : "button";
