+++ app/components/ui/searchBar.js
@@ -5,6 +5,14 @@ import { FaSearch } from 'react-icons/fa';
 import { XCircle, Timer, Coffee, Utensils } from 'lucide-react';
 import { useRouter } from 'next/navigation';
 
+/**
+ * Highlights search terms in a given text by wrapping the matching parts with a styled span.
+ *
+ * @param {Object} props - Component props.
+ * @param {string} props.text - The text to be searched for highlights.
+ * @param {Array<string>} props.searchTerms - The terms to highlight in the text.
+ * @returns {JSX.Element} - The text with highlighted search terms.
+ */
 const HighlightedText = ({ text, searchTerms }) => {
     if (!searchTerms || searchTerms.length === 0) return text;
 
@@ -31,6 +39,17 @@ const HighlightedText = ({ text, searchTerms }) => {
     );
 };
 
+/**
+ * RecipeSearchBar component that provides a search input with real-time suggestions for recipes.
+ * It fetches recipe suggestions based on the user's search input and allows users to search by title, ingredient, or category.
+ * It also displays recent searches and allows users to clear them.
+ *
+ * @param {Object} props - Component props.
+ * @param {string} [props.currentSearch] - The initial search query.
+ * @param {Function} props.onSearch - A callback function that is called when a search is made.
+ * @param {number} [props.minCharacters=2] - The minimum number of characters required to trigger search suggestions.
+ * @returns {JSX.Element} - The RecipeSearchBar component with input field and suggestion list.
+ */
 const RecipeSearchBar = ({
     currentSearch = '',
     onSearch,
@@ -43,12 +62,18 @@ const RecipeSearchBar = ({
     const [recentSearches, setRecentSearches] = useState([]);
     const [selectedRecipe, setSelectedRecipe] = useState(null);
     const router = useRouter();
-  
+
 
     useEffect(() => {
         setSearch(currentSearch);
     }, [currentSearch]);
 
+    /**
+     * Formats the search input into individual terms to be used in the search query.
+     *
+     * @param {string} searchText - The search input text.
+     * @returns {Array<string>} - An array of search terms.
+     */
     const getSearchTerms = (searchText) => {
         return searchText
             .toLowerCase()
@@ -56,6 +81,11 @@ const RecipeSearchBar = ({
             .filter(term => term.length >= minCharacters);
     };
 
+    /**
+     * Saves the search term to local storage as a recent search.
+     *
+     * @param {string} searchTerm - The search term to be saved.
+     */
     const saveRecentSearch = (searchTerm) => {
         const updatedSearches = [
             searchTerm,
@@ -66,37 +96,40 @@ const RecipeSearchBar = ({
         localStorage.setItem('recentRecipeSearches', JSON.stringify(updatedSearches));
     };
 
+    /**
+     * Fetches recipe suggestions based on the current search input.
+     */
     useEffect(() => {
         const fetchSuggestions = async () => {
             if (search.length < minCharacters) {
                 setSuggestions([]);
                 return;
             }
-        
+
             try {
                 setLoading(true);
-        
+
                 // Check for an exact match
                 const exactMatchResponse = await fetch(`/api/recipes?exactTitle=${encodeURIComponent(search)}`);
                 if (!exactMatchResponse.ok) throw new Error('Failed to check for exact match');
                 const exactMatchData = await exactMatchResponse.json();
-        
+
                 if (exactMatchData.recipes && exactMatchData.recipes.length > 0) {
                     setSelectedRecipe(exactMatchData.recipes[0]);
                     setShowSuggestions(false);
                     return;
                 }
-        
+
                 // Fallback to general suggestions
                 const params = new URLSearchParams({
                     search,
                     limit: '10',
                     page: '1',
                 });
-        
+
                 const response = await fetch(`/api/recipes?${params}`);
                 if (!response.ok) throw new Error('Failed to fetch suggestions');
-        
+
                 const data = await response.json();
                 setSuggestions(data.recipes);
             } catch (error) {
@@ -106,18 +139,22 @@ const RecipeSearchBar = ({
                 setLoading(false);
             }
         };
-        
-    
+
+
         const timeoutId = setTimeout(fetchSuggestions, 500);
         return () => clearTimeout(timeoutId);
     }, [search, minCharacters]);
-    
-    
+
+    /**
+     * Fetches details of a specific recipe by ID.
+     *
+     * @param {string} recipeId - The ID of the recipe to fetch.
+     */
     const fetchRecipeDetails = async (recipeId) => {
         try {
             const response = await fetch(`/api/recipes/${recipeId}`);
             if (!response.ok) throw new Error('Failed to fetch recipe details');
-            
+
             const data = await response.json();
             setSelectedRecipe(data.recipe);
         } catch (error) {
@@ -125,6 +162,11 @@ const RecipeSearchBar = ({
         }
     };
 
+    /**
+     * Handles the form submission for the search input.
+     *
+     * @param {Event} e - The form submit event.
+     */
     const handleSearchSubmit = (e) => {
         e.preventDefault();
         if (search.trim()) {
@@ -139,6 +181,11 @@ const RecipeSearchBar = ({
         }
     };
 
+    /**
+     * Handles the click on a suggestion.
+     *
+     * @param {Object} suggestion - The selected suggestion object.
+     */
     const handleSuggestionClick = (suggestion) => {
         setSearch(suggestion.title);
         saveRecentSearch(suggestion.title);
@@ -147,16 +194,22 @@ const RecipeSearchBar = ({
         fetchRecipeDetails(suggestion._id);
     };
 
+    /**
+     * Clears the current search input and suggestions.
+     */
     const clearSearch = () => {
         setSearch('');
         setShowSuggestions(false);
         setSuggestions([]);
-        setSelectedRecipe(null); 
+        setSelectedRecipe(null);
         if (onSearch) {
             onSearch('');
         }
     };
 
+    /**
+     * Clears the recent search history stored in local storage.
+     */
     const clearRecentSearches = () => {
         setRecentSearches([]);
         localStorage.removeItem('recentRecipeSearches');
@@ -168,9 +221,9 @@ const RecipeSearchBar = ({
         <div className="max-w-3xl mx-auto">
             <form onSubmit={handleSearchSubmit} className="relative">
                 <div className="relative flex items-center">
-                <div className="absolute left-4 top-1/2 transform -translate-y-1/2 text-[#ff4f1a]"> 
+                    <div className="absolute left-4 top-1/2 transform -translate-y-1/2 text-[#ff4f1a]">
                         <FaSearch className="h-4 w-4" />
-                        </div>
+                    </div>
                     <input
                         type="text"
                         value={search}
@@ -194,7 +247,7 @@ const RecipeSearchBar = ({
                                 <XCircle className="h-5 w-5" />
                             </button>
                         )}
-                       {/* <button 
+                        {/* <button 
                             type="submit" 
                             className=" text-gray-400 p-3 "
                             aria-label="Search"
@@ -202,10 +255,10 @@ const RecipeSearchBar = ({
                         </button>*/}
                     </div>
                 </div>
-    
+
                 {showSuggestions && (
-                    <div 
-                         className="absolute z-10 w-full mt-2 bg-white rounded-xl shadow-lg border border-gray-100 overflow-hidden"
+                    <div
+                        className="absolute z-10 w-full mt-2 bg-white rounded-xl shadow-lg border border-gray-100 overflow-hidden"
                         onMouseDown={(e) => e.preventDefault()}
                     >
                         {loading ? (
@@ -216,7 +269,7 @@ const RecipeSearchBar = ({
                         ) : suggestions.length > 0 ? (
                             <ul className="divide-y divide-gray-100">
                                 {suggestions.map((suggestion, index) => (
-                                    <li 
+                                    <li
                                         key={suggestion._id || index}
                                         className="hover:bg-gray-50 transition-colors cursor-pointer"
                                         onClick={() => handleSuggestionClick(suggestion)}
@@ -225,7 +278,7 @@ const RecipeSearchBar = ({
                                             <div className="flex items-center mb-2">
                                                 <Utensils className="h-4 w-4 text-teal-500 mr-3" />
                                                 <span className="font-medium text-gray-800">
-                                                    <HighlightedText 
+                                                    <HighlightedText
                                                         text={suggestion.title}
                                                         searchTerms={searchTerms}
                                                     />
@@ -244,8 +297,8 @@ const RecipeSearchBar = ({
                         ) : search.length >= minCharacters ? (
                             <div className="p-6 text-gray-600 text-center font-medium">
                                 No recipes found matching &quot;{search}&quot;
-                                <button 
-                                    onClick={clearRecentSearches} 
+                                <button
+                                    onClick={clearRecentSearches}
                                     className="mt-2 text-green-600 hover:underline"
                                 >
                                     Clear recent searches
@@ -256,7 +309,7 @@ const RecipeSearchBar = ({
                                 <h3 className="text-sm font-medium text-gray-400 uppercase tracking-wider mb-3 px-2">Recent Searches</h3>
                                 <ul className="space-y-2">
                                     {recentSearches.map((term, index) => (
-                                        <li 
+                                        <li
                                             key={index}
                                             className="cursor-pointer text-gray-700 hover:text-teal-500 flex items-center p-2 rounded-lg hover:bg-gray-50 transition-colors duration-200"
                                             onClick={() => {
@@ -275,7 +328,7 @@ const RecipeSearchBar = ({
                     </div>
                 )}
             </form>
-    
+
             {selectedRecipe && (
                 <div className="mt-8 p-6 bg-white rounded-xl shadow-md border border-gray-100">
                     <h2 className="text-2xl font-bold text-gray-800 mb-3">{selectedRecipe.title}</h2>
@@ -288,7 +341,7 @@ const RecipeSearchBar = ({
             )}
         </div>
     );
-    
+
 };
 
 export default RecipeSearchBar;
