+++ app/components/voice-ai/RecordVoice.js
@@ -2,11 +2,21 @@
 
 import React, { useEffect, useState } from 'react';
 
+/**
+ * RecordVoice component - Allows users to follow instructions step-by-step with voice feedback
+ * and control using voice commands.
+ * 
+ * @param {Object} props - The component props.
+ * @param {Array} props.instructions - List of instructions to be read aloud.
+ * @returns {JSX.Element} The RecordVoice component.
+ */
 const RecordVoice = ({ instructions = [] }) => {
     const [currentStep, setCurrentStep] = useState(0); // Track the current instruction step
     const [isListening, setIsListening] = useState(false); // Track listening status
 
-    // Function to read the current instruction
+    /**
+     * Reads the current instruction aloud using speech synthesis.
+     */
     const readInstruction = () => {
         if (currentStep < instructions.length) {
             const message = new SpeechSynthesisUtterance(instructions[currentStep]);
@@ -17,7 +27,9 @@ const RecordVoice = ({ instructions = [] }) => {
         }
     };
 
-    // Function to handle speech recognition
+    /**
+     * Handles the voice commands for navigation between instructions.
+     */
     const handleVoiceCommands = () => {
         const recognition = new (window.SpeechRecognition || window.webkitSpeechRecognition)();
         recognition.lang = 'en-US'; // Set language
